fixing NullPointer and wrong type of ModelValue.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/Princess@12638 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/predicates/princess/PrincessModel.java
@@ -147,7 +147,7 @@ class PrincessModel {
         break;
 
       case Integer:
-        lValue = ((SimpleAPI.IntValue)lValueTerm).v();
+        lValue = ((SimpleAPI.IntValue)lValueTerm).v().longValue();
         break;
 
       default:
